@@ -1,4 +1,4 @@
 package gomongostreams
 
 // Version of the package
-const Version = "0.1.0"
+const Version = "2.0.0"
